requirements.txt
PyQt6>=6.4.0
Pillow>=10.0.0
torch
torchvision
transformers==4.43.0
accelerate
